ai/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
colorlog
requests
openai-agents
mcp
pytest 
pytest-asyncio
google-api-python-client